chore: extract getTaskNames
